Created CircleImageWithTextInput, updated CircleImage and TextInput #17
Created the CircleImage WithTextInput molecule. When creating it I found out that the Circle Image and the Text Input were not the same height. So I updated them to be the same height. Whilst using the comment variant of the text input I found out that it was trying to display an image with that variant. The TextInput atom was not responsible for display images. Using multiple atoms together is the molecule's responsibility. So I removed the unneeded code from the TextInput component. Circle image component took in a prop of clickAction but never used it. So I added it to be fired onClick of the function.